moving PointLight and PointLightShadowMap to point_light.rs
