+++ crates/bevy_pbr/src/light/mod.rs
@@ -22,6 +22,8 @@ use crate::*;
 
 mod ambient_light;
 pub use ambient_light::AmbientLight;
+mod point_light;
+pub use point_light::{PointLight, PointLightShadowMap};
 
 /// Constants for operating with the light units: lumens, and lux.
 pub mod light_consts {
@@ -77,73 +79,6 @@ pub mod light_consts {
         pub const FULL_DAYLIGHT: f32 = 20_000.;
         /// The amount of light (lux) in direct sunlight.
         pub const DIRECT_SUNLIGHT: f32 = 100_000.;
-    }
-}
-
-/// A light that emits light in all directions from a central point.
-///
-/// Real-world values for `intensity` (luminous power in lumens) based on the electrical power
-/// consumption of the type of real-world light are:
-///
-/// | Luminous Power (lumen) (i.e. the intensity member) | Incandescent non-halogen (Watts) | Incandescent halogen (Watts) | Compact fluorescent (Watts) | LED (Watts |
-/// |------|-----|----|--------|-------|
-/// | 200  | 25  |    | 3-5    | 3     |
-/// | 450  | 40  | 29 | 9-11   | 5-8   |
-/// | 800  | 60  |    | 13-15  | 8-12  |
-/// | 1100 | 75  | 53 | 18-20  | 10-16 |
-/// | 1600 | 100 | 72 | 24-28  | 14-17 |
-/// | 2400 | 150 |    | 30-52  | 24-30 |
-/// | 3100 | 200 |    | 49-75  | 32    |
-/// | 4000 | 300 |    | 75-100 | 40.5  |
-///
-/// Source: [Wikipedia](https://en.wikipedia.org/wiki/Lumen_(unit)#Lighting)
-#[derive(Component, Debug, Clone, Copy, Reflect)]
-#[reflect(Component, Default)]
-pub struct PointLight {
-    pub color: Color,
-    /// Luminous power in lumens, representing the amount of light emitted by this source in all directions.
-    pub intensity: f32,
-    pub range: f32,
-    pub radius: f32,
-    pub shadows_enabled: bool,
-    pub shadow_depth_bias: f32,
-    /// A bias applied along the direction of the fragment's surface normal. It is scaled to the
-    /// shadow map's texel size so that it can be small close to the camera and gets larger further
-    /// away.
-    pub shadow_normal_bias: f32,
-}
-
-impl Default for PointLight {
-    fn default() -> Self {
-        PointLight {
-            color: Color::WHITE,
-            // 1,000,000 lumens is a very large "cinema light" capable of registering brightly at Bevy's
-            // default "very overcast day" exposure level. For "indoor lighting" with a lower exposure,
-            // this would be way too bright.
-            intensity: 1_000_000.0,
-            range: 20.0,
-            radius: 0.0,
-            shadows_enabled: false,
-            shadow_depth_bias: Self::DEFAULT_SHADOW_DEPTH_BIAS,
-            shadow_normal_bias: Self::DEFAULT_SHADOW_NORMAL_BIAS,
-        }
-    }
-}
-
-impl PointLight {
-    pub const DEFAULT_SHADOW_DEPTH_BIAS: f32 = 0.02;
-    pub const DEFAULT_SHADOW_NORMAL_BIAS: f32 = 0.6;
-}
-
-#[derive(Resource, Clone, Debug, Reflect)]
-#[reflect(Resource)]
-pub struct PointLightShadowMap {
-    pub size: usize,
-}
-
-impl Default for PointLightShadowMap {
-    fn default() -> Self {
-        Self { size: 1024 }
     }
 }
